Backend/requirements.txt
imgaug==0.4.0
labelme==5.0.5
numpy==1.23.4
opencv-python==4.7.0.68
pillow==9.3.0
pyqt5==5.15.7
pytorch-lightning==1.8.0.post1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm==4.64.1
